+++ docs/storybook/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.24.7](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.24.6...@docs/storybook@3.24.7) (2026-05-04)
+
+**Note:** Version bump only for package @docs/storybook
+
 ## [3.24.6](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.24.5...@docs/storybook@3.24.6) (2026-04-29)
 
 **Note:** Version bump only for package @docs/storybook