+++ terezinha-farm/vite-app/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.31](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.17.30...@terezinha-farm/vite-app@0.17.31) (2026-05-04)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
 ## [0.17.30](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.17.29...@terezinha-farm/vite-app@0.17.30) (2026-04-28)
 
 **Note:** Version bump only for package @terezinha-farm/vite-app